+++ frontend/app/components/quest/TaskListItem.vue
@@ -1,0 +1,62 @@
+<template>
+  <div class="d-flex align-items-center">
+    <i class="fa fa-circle-o fa-lg"></i>
+    <input
+      v-model="task.name"
+      class="focusable"
+      :class="isSubtask ? 'subtask-name' : 'task-name'"
+      placeholder="タスクを入力"
+      @blur="updateTask(task)"
+      @keydown.enter="addNewTask"
+      @keydown.prevent.down="moveNext"
+      @keydown.prevent.up="movePrev"
+    />
+    <b-badge v-if="!isSubtask" variant="primary" pill class="ml-auto">{{
+      task.subtasks.length
+    }}</b-badge>
+  </div>
+</template>
+
+<script>
+export default {
+  props: ['task', 'isSubtask'],
+  methods: {
+    updateTask() {
+      this.$emit('updateTask', this.task);
+    },
+    addNewTask(e) {
+      if (e.keyCode !== 13) return; // 日本語入力確定を除外
+      this.$emit('addNewTask', e, this.task.id);
+    },
+    moveNext(event) {
+      const elements = document.getElementsByClassName('focusable');
+      const index = [].findIndex.call(elements, (e) => e === event.target);
+      if (index + 1 < elements.length) elements[index + 1].focus();
+    },
+    movePrev(event) {
+      const elements = document.getElementsByClassName('focusable');
+      const index = [].findIndex.call(elements, (e) => e === event.target);
+      if (index - 1 >= 0) elements[index - 1].focus();
+    },
+  },
+};
+</script>
+
+
+<style scoped>
+*:focus {
+  outline: none;
+}
+
+.task-name,
+.subtask-name {
+  width: 100%;
+  border: none;
+  background: none;
+  margin-left: 1em;
+}
+
+.task-name {
+  font-weight: bold;
+}
+</style>
